fixed fullscreendialog sizing
