+++ app/src/main/java/io/treehouses/remote/bases/FullScreenDialogFragment.kt
@@ -1,0 +1,17 @@
+package io.treehouses.remote.bases
+
+import android.app.Dialog
+import android.os.Bundle
+import android.view.ViewGroup
+
+open class FullScreenDialogFragment : BaseDialogFragment() {
+    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
+        val dialog: Dialog = super.onCreateDialog(savedInstanceState)
+        return dialog
+    }
+
+    override fun onStart() {
+        super.onStart()
+        dialog?.window?.setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
+    }
+}
